+++ src/db/sqlite/sql/authentication-update-credential.sql
@@ -1,0 +1,3 @@
+UPDATE authentication
+	SET credential = :credential
+	WHERE identifier = :identifier